// src/app/models/contrato.model.ts
export class contratoModel {

    _id: string;
    code: string;
    date: Date;
    company: string;
    employee: string;
    vehicle: string;
    customer: string;
    financialData: string;
    legalRepresentative: string;
    finantialInstitution: string;
    ruc: string;
    endoso: string;

    constructor() {}

}